requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
fastapi
h11==0.16.0
httpcore==1.0.9
idna==3.10
jiter==0.10.0
lxml==5.4.0
numpy==2.2.6
openai
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pydantic==2.11.5
pydantic_core==2.33.2
PyMuPDF==1.26.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
six==1.17.0
sniffio==1.3.1
sqlalchemy
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
XlsxWriter==3.2.3
